Adjust primary login chain
When integrating the original auth function with my code, the first cypress test run would end up redirecting to the standard auth flow rather than ROPC.  The second test run would use the desired ROPC route.  This change customizes the flow direction based on the existence of the cached token.  If a non-expired token is passed to the login() function, we simply navigate to the root page and return the same token back to the caller.  If no token is passed, then we initiate the ROPC flow, process the token result and then navigate to the root page and return back the new token.